Vision_agent/pyproject.toml
[project]
name = "vision-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.26",
    "adk-redis[memory]>=0.0.4",
    "asyncpg>=0.31.0",
    "google-adk-redis>=0.1.5",
    "google-adk[a2a]>=1.31.0",
    "google-cloud-aiplatform>=1.148.0",
    "google-cloud-documentai>=3.14.0",
    "google-cloud-firestore>=2.27.0",
    "google-cloud-storage>=3.10.1",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "numpy>=2.4.4",
    "opencv-python-headless>=4.13.0.92",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-gcp-trace>=1.11.0",
    "opentelemetry-instrumentation-google-genai>=0.7b0",
    "opentelemetry-instrumentation-requests>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
    "redisvl>=0.17.0",
]
